Move ObserveMaxWidth into WebView
